amend: rename draggable to drag
