+++ packages/react-dataparcels-draggable/README.md
@@ -1,6 +1,6 @@
 ![dataparcels](https://user-images.githubusercontent.com/345320/48319791-4eece200-e666-11e8-8b19-252cd1135ae2.png)
 
-<a href="https://www.npmjs.com/package/react-dataparcels-draggable"><img src="https://img.shields.io/npm/v/react-dataparcels-draggable.svg?style=flat-square"></a>
+<a href="https://www.npmjs.com/package/react-dataparcels-drag"><img src="https://img.shields.io/npm/v/react-dataparcels-drag.svg?style=flat-square"></a>
 [![CircleCI](https://circleci.com/gh/blueflag/dataparcels/tree/master.svg?style=shield)](https://circleci.com/gh/blueflag/dataparcels/tree/master)
 
 A plugin for [`react-dataparcels`](https://www.npmjs.com/package/react-dataparcels) that adds drag and drop re-ordering of elements, using the wonderful [react-sortable-hoc](https://github.com/clauderic/react-sortable-hoc). 
@@ -11,4 +11,4 @@ A plugin for [`react-dataparcels`](https://www.npmjs.com/package/react-dataparce
 
 ## Packages
 
-Get [`react-dataparcels-draggable`](https://www.npmjs.com/package/react-dataparcels-draggable).
+Get [`react-dataparcels-drag`](https://www.npmjs.com/package/react-dataparcels-drag).
