+++ packages/react-dataparcels-draggable/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "react-dataparcels-draggable",
+  "name": "react-dataparcels-drag",
   "version": "0.18.0-2",
   "description": "A plugin for react-dataparcels that adds drag and drop re-ordering of elements.",
   "main": "lib/index.js",
